revert: harden MiniCPM multimodal integration
This reverts commit 84df79c7f8fa73369508dc111d16d9d8de4593d0.
