+++ areno/accel/ops.py
@@ -2,7 +2,7 @@
 
 Re-exports fused activation wrappers and the Triton-based kernels
 (fused MoE experts, grouped RMSNorm with sigmoid gate, segmented linear
-attention). Also exposes small utilities used throughout the layer code:
+attention). Adds two small utilities used throughout the layer code:
 
 - ``log_once`` / ``warn_once``: emit a logger message exactly once per
   process for a given key, so kernel-selection diagnostics do not flood
@@ -12,6 +12,7 @@ attention). Also exposes small utilities used throughout the layer code:
 
 from __future__ import annotations
 
+import logging
 from typing import Any
 
 import torch
@@ -27,7 +28,25 @@ from areno.accel.kernels.fused_moe import fused_experts as areno_fused_experts
 from areno.accel.kernels.fused_moe import is_available as fused_moe_is_available
 from areno.accel.kernels.group_rmsnorm import rms_norm_gate_fwd
 from areno.accel.kernels.seg_la import SegLaMeta, seg_la_fwd
-from areno.engine.log import log_once, warn_once
+
+logger = logging.getLogger(__name__)
+# Process-wide set of message keys already emitted by log_once/warn_once.
+_LOGGED: set[str] = set()
+
+
+def log_once(key: str, message: str, *, level: int = logging.DEBUG) -> None:
+    """Log ``message`` at most once per process for the given ``key``."""
+
+    if key in _LOGGED:
+        return
+    logger.log(level, message)
+    _LOGGED.add(key)
+
+
+def warn_once(key: str, message: str) -> None:
+    """Emit a warning at most once per process for the given ``key``."""
+
+    log_once(key, message, level=logging.WARNING)
 
 
 @torch._dynamo.disable
